Combine original FilterBox with modal version
