+++ client/src/pages/Feed.js
@@ -2,7 +2,6 @@ import React from "react";
 import styled from "styled-components";
 
 import FilterBox from "../components/FilterBox/FilterBox";
-import FilterBoxModal from "../components/FilterBox/FilterBoxModal";
 
 const FeedWraper = styled.div`
   width: 100%;
@@ -13,7 +12,6 @@ const Feed = () => {
   return (
     <FeedWraper>
       <FilterBox />
-      <FilterBoxModal />
     </FeedWraper>
   );
 };
